+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - Timed attribute to automatically show how long a method or property takes to execute.
 - MarkInProfiler attribute to make methods show up in the Unity profiler.
+- Added %class_full%, %method_full%, and %property_full% for messages.
 
 ### Changed
 - ResetStatic no longer needs default values.